Completed and Tested

+++ Mvc_ShoppingCart/Mvc_ShoppingCart/Controllers/CartController.cs
@@ -9,6 +9,7 @@ using System.IO;
 using System.Security.Cryptography;
 using DataAccess;
 using System.Text;
+using Mvc_ShoppingCart.Models;
 
 namespace Mvc_ShoppingCart.Controllers
 {
@@ -28,7 +29,7 @@ namespace Mvc_ShoppingCart.Controllers
         /// <param name="qty">qty bought</param>
         /// <returns>view</returns>
         /// 
-        [Authorize]
+        [CustomAuthorize()]
         public ActionResult AddToShoppingCart(int Quantity = 1)
         {
 
@@ -46,6 +47,8 @@ namespace Mvc_ShoppingCart.Controllers
             {
                 List<Order> list = new ShoppingCartBL().GetOrdersByUsername(User.Identity.Name).ToList();
                 List<OrderDetail> odList = new List<OrderDetail>();
+                List<ShoppingCart> shoppingCartProducts =  new List<ShoppingCart>();
+                shoppingCartProducts = new ShoppingCartBL().GetShoppingCartItemsByUsername(User.Identity.Name).ToList();
 
                 foreach (Order o in list)
                 {
@@ -57,6 +60,16 @@ namespace Mvc_ShoppingCart.Controllers
                 sc.Qty = Quantity;
                 sc.ProductID = Convert.ToInt32(Session["ProductID"]);
 
+                //chek if item is already in shopping cart
+                foreach (ShoppingCart ss in shoppingCartProducts)
+                {
+                    if (ss.ProductID == sc.ProductID)
+                    {
+                        ViewBag.Msg = "Product Is already in Shopping Cart";
+                        return View();
+                    }
+                }
+                //check if item is bought
                 foreach (OrderDetail od in odList)
                 {
                     if (od.ProductID == sc.ProductID)
@@ -143,6 +156,12 @@ namespace Mvc_ShoppingCart.Controllers
 
         //
         // GET: /Cart/Delete/5
+        public ActionResult Error()
+        {
+
+            return View() ;
+        }
+
 
         public ActionResult Delete(string Username, int ProductID)
         {
@@ -152,7 +171,7 @@ namespace Mvc_ShoppingCart.Controllers
 
         public ActionResult DownloadFile(string Username, int ProductID)
         {
-            Product p = new ProductBL().GetProductsByID(ProductID);
+            Common.Product p = new ProductBL().GetProductsByID(ProductID);
             User u = new UserBL().GetUserByEmail(p.SellerEmail);
             string pathIn = Server.MapPath(p.ZipPath.ToString());
             string pathOut = Server.MapPath(p.ZipPath.Replace("Enc.","."));
@@ -161,21 +180,27 @@ namespace Mvc_ShoppingCart.Controllers
             byte[] IV = he.AsymmetricDecryption(u.PrivateKey, p.IV);
 
             MemoryStream ms = he.Decryption(pathIn, pathOut, Key, IV, ProductID);
+            if (ms == null)
+            {
+                return RedirectToAction("Error");
+            }
+
             System.IO.File.Delete(pathOut);
             try
             {
-                string filename = pathOut.Substring(pathOut.LastIndexOf("\\")).Remove(0,1);
+                string filename = pathOut.Substring(pathOut.LastIndexOf("\\")).Remove(0, 1);
                 byte[] bytesInStream = ms.ToArray();
                 Response.Clear();
                 Response.ContentType = "application/force-download";
-                Response.AddHeader("content-disposition", "attachment;    filename="+filename+"");
+                Response.AddHeader("content-disposition", "attachment;    filename=" + filename + "");
                 Response.BinaryWrite(bytesInStream);
                 Response.End();
             }
             catch (Exception e)
             {
-               
+
             }
+            
 
             return RedirectToAction("Index");
         }
